pyproject.toml
[tool.poetry]
name = "asf-v3"
version = "0.1.0"
description = ""
authors = ["anthonymanciny <141266961+anthonymanciny@users.noreply.github.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.6"
uvicorn = "^0.32.1"
sqlalchemy = "^2.0.36"
psycopg2-binary = "^2.9.10"
pytest = "^8.3.4"
alembic = "^1.14.0"
python-decouple = "^3.8"
httpx = "^0.28.0"
passlib = "^1.7.4"
pymysql = "^1.1.1"
python-jose = "^3.3.0"
python-multipart = "^0.0.19"
python-dotenv = "^1.0.1"
mariadb = "^1.1.11"
bcrypt = "^4.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
